CORS validation added

+++ index.js
@@ -6,10 +6,16 @@ const express = require('express'),
       Models = require ('./models'),
       Movies = Models.Movie,
       Users = Models.User;
-     		
+     
+// middleware for parsing requests      
 app.use(express.json());
 app.use(express.urlencoded({extended:true}));
 app.use(express.static('public'));
+
+//validation
+const cors = require ('cors');
+app.use(cors());
+const { check, validationResult } = require('express-validator');
 
 //Authentication
 let auth = require('./auth')(app);
@@ -19,7 +25,15 @@ require('./passport');
 //Log requests to server
 app.use(morgan('common'));
 
+//MongoDB connection via Mongoose
 mongoose.connect('mongodb://localhost:27017/cfDB');
+
+//Port designaiton
+const port = process.env.PORT || 8080;
+app.listen(port, '0.0.0.0',() => {
+ console.log('Listening on Port ' + port);
+});
+
 
 //Default text response when at /
 app.get('/', (req, res) =>{
@@ -76,8 +90,23 @@ app.get('/movies/directors/:directorName', passport.authenticate('jwt', { sessio
 });
 
 
-//Add a new user
-app.post('/users', async (req, res) => {
+//Add a new user with validation logic for request
+app.post('/signup', [
+    check('Username', 'Username is required').isLength({min: 5}),
+    check('Username', 'Username contains non alphanumeric characters - not allowed.').isAlphanumeric(),
+    check('Password', 'Password is required').not().isEmpty(),
+    check('Email', 'Email does not appear to be valid').isEmail()
+  ], async (req, res) => {
+
+  // check the validation object for errors
+    let errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+      return res.status(422).json({ errors: errors.array() });
+    }
+
+  let hashedPassword = Users.hashPassword(req.body.Password);
+
   await Users.findOne({ Username: req.body.Username })
     .then((user) => {
       if (user) {
@@ -86,15 +115,15 @@ app.post('/users', async (req, res) => {
         Users
           .create({
             Username: req.body.Username,
-            Password: req.body.Password,
+            Password: hashedPassword,
             Email: req.body.Email,
             Birthday: req.body.Birthday,
-				})
+          })
           .then((user) =>{res.status(201).json(user) })  
           .catch((error) => {
             console.error(error);
             res.status(500).send('Error: ' + error);
-          })      	
+          });      	
       }
     })
     .catch((err) => {
@@ -103,13 +132,22 @@ app.post('/users', async (req, res) => {
     });
 });
 
+
 //Allow users to update their user info 
-app.put('/users/:Username', passport.authenticate('jwt', { session: false }), async (req, res) => {
-   // CONDITION TO CHECK ADDED HERE
-   if(req.user.Username !== req.params.Username){
+app.put('/users/:Username', [
+  check('Username', 'Username is required').isLength({min: 5}),
+  check('Username', 'Username contains non alphanumeric characters - not allowed.').isAlphanumeric(),
+  check('Password', 'Password is required').not().isEmpty(),
+  check('Email', 'Email does not appear to be valid').isEmail()
+],
+passport.authenticate('jwt', { session: false }), async (req, res) => {
+  let errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }
+  if(req.user.Username !== req.params.Username){
     return res.status(400).send('Permission denied');
-}
-// CONDITION ENDS  
+  }
   await Users.findOneAndUpdate({ Username: req.params.Username }, { $set:
     {
       Username: req.body.Username,
@@ -131,10 +169,15 @@ app.put('/users/:Username', passport.authenticate('jwt', { session: false }), as
 
 //Allow users to add a movie to their list of favorites
 app.post('/users/:Username/Movies/:MovieID', passport.authenticate('jwt', { session: false }), async (req, res) => {
+  // CONDITION TO CHECK ADDED HERE
+  if(req.user.Username !== req.params.Username){
+    return res.status(400).send('Permission denied');}
+  // CONDITION ENDS  
+
   await Users.findOneAndUpdate({ Username: req.params.Username }, {
      $push: { FavoriteMovies: req.params.MovieID }
    },
-   { new: true }) // This line makes sure that the updated document is returned
+   { new: true }) 
   .then((updatedUser) => {
     res.json(updatedUser);
   })
@@ -147,22 +190,32 @@ app.post('/users/:Username/Movies/:MovieID', passport.authenticate('jwt', { sess
 
 //Allow users to remove a movie from their list of favorites
 app.delete('/users/:Username/Movies/:MovieID', passport.authenticate('jwt', { session: false }), async (req, res) => {
-    await Users.findOneAndUpdate({ Username: req.params.Username },	{ 
+  // CONDITION TO CHECK ADDED HERE
+  if(req.user.Username !== req.params.Username){
+    return res.status(400).send('Permission denied');}
+  // CONDITION ENDS  
+
+  await Users.findOneAndUpdate({ Username: req.params.Username },	{ 
     $pull: { FavoriteMovies: req.params.MovieID } 
-    },
-    { new: true }) // This line makes sure that the updated document is returned
-    .then ((updatedUser) => {
-      res.json(updatedUser);
-    })
-    .catch((err) => {
-      console.error(err);
-      res.status(500).send('Error: ' + err);
-    });
+  },
+  { new: true })
+  .then ((updatedUser) => {
+    res.json(updatedUser);
+  })
+  .catch((err) => {
+    console.error(err);
+    res.status(500).send('Error: ' + err);
+  });
 });
 
 
 //Allow existing users to deregister
 app.delete('/users/:Username', passport.authenticate('jwt', { session: false }), async (req, res) => {
+  // CONDITION TO CHECK ADDED HERE
+  if(req.user.Username !== req.params.Username){
+    return res.status(400).send('Permission denied');}
+  // CONDITION ENDS   
+  
   await Users.findOneAndDelete({ Username: req.params.Username })
     .then((user) => {
       if (!user) {
@@ -177,5 +230,15 @@ app.delete('/users/:Username', passport.authenticate('jwt', { session: false }),
     });
 });
 
+//Return a list of all users
+app.get('/users', async (req, res) => {
+	await Users.find()
+	.then((users) => {
+		res.status(201).json(users);
+	})
+	.catch((err) => {
+		console.error(err);
+		res.status(500).send('Error: ' + err);
+	});
+})
 
-app.listen(8080, () => console.log('Listening on port 8080.'));
